Test to verify sriov physical function neutron port assignment
This test verifies that instances can be booted successfully
with direct assignment of physical functions. After the
instance is created, the pci device information that is tied
to the binding profile of the neutron port and that present
in the nova database are checked for consistency.

DocImpact
Verifies: blueprint sriov-pf-passthrough-neutron-port